requirements.txt
Django>=4.2
djangorestframework>=3.14
djongo>=1.4.1
pymongo>=4.0
djangorestframework-simplejwt>=5.2
python-dateutil
